requirements.txt
Flask
requests
beautifulsoup4
python-dotenv
supabase
google-genai
tqdm
gunicorn
markdown
google-cloud-storage
pandas
numpy